Added delay after initial triggering
After airplane goes through, added code to wait 20ms and then read the
sensor again. This is only 1/50 sec and should still catch all paper
airplanes, since there was code to check every 1/20th sec before and it
worked almost every time, but 1/50sec should be enough time to make sure
there's actually a paper airplane in the target, and not just a
momentary fluke of noise on the signal line or who knows what.